refactor: [SSP-13] success noty and route
