+++ app/Http/Controllers/Admin/ProductController.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Admin;
 
 use App\Enums\Status;
 use App\Models\Category;
